requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-genai

# News Scraping
playwright==1.48.0
newspaper3k==0.2.8
beautifulsoup4==4.12.3
lxml==5.3.0
lxml_html_clean==0.4.3
googlenewsdecoder==0.1.7
selectolax==0.4.4
feedparser==6.0.11
requests==2.32.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Optional: Scheduling
APScheduler==3.10.4

# Sentiment Analysis
transformers>=4.35.0
torch>=2.0.0
